use proper structure for the UI
