+++ templates/plugin-template/src/ui/app/App.tsx
