+++ templates/plugin-template/src/ui/app/constants.ts
