+++ templates/plugin-template/src/ui/app/features/content-objects/index.ts
@@ -1,0 +1,2 @@
+export { ContentObjectsView } from './ContentObjectsView';
+export type { SortField, FilterableField } from './types';
