+++ templates/plugin-template/src/ui/app/layouts/AppSidebarItem.tsx
