+++ templates/plugin-template/src/ui/app/layouts/OrgGate.tsx
