+++ templates/plugin-template/src/ui/app/layouts/PluginLayout.tsx
