+++ templates/plugin-template/src/ui/app/layouts/PluginTopNav.tsx
