+++ templates/plugin-template/src/ui/app/pages/ContentObjectsPage.tsx
@@ -1,0 +1,5 @@
+import { ContentObjectsView } from '../features/content-objects';
+
+export function ContentObjectsPage() {
+    return <ContentObjectsView />;
+}
