+++ templates/plugin-template/src/ui/main.tsx
@@ -7,12 +7,12 @@ import './index.css'
 // initialize dev environment
 import { AdminApp } from '@vertesia/tools-admin-ui'
 import { RouterProvider, type Route } from '@vertesia/ui/router'
-import { App } from './app'
+import { App } from './app/App'
 import { setUsePluginAssets } from './assets'
 import "./env"
-import { OrgGate } from './OrgGate'
-import { PluginAccessDenied } from './PluginAccessDenied'
-import { PluginLayout } from './PluginLayout'
+import { OrgGate } from './app/layouts/OrgGate'
+import { PluginAccessDenied } from './app/layouts/PluginAccessDenied'
+import { PluginLayout } from './app/layouts/PluginLayout'
 
 setUsePluginAssets(false);
